+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-04-11 23:03:22
-// Commit: fix(security): prevent double tap duplicate action in security
+// Updated at: 2026-04-12 16:54:14
+// Commit: docs(security): update readme and API contracts for security
 
 class securityComponentState {
-  final String commitRef = '323';
-  final String lastMessage = 'fix(security): prevent double tap duplicate action in security';
-  final DateTime timestamp = DateTime.parse('2026-04-11T23:03:22+05:00');
+  final String commitRef = '351';
+  final String lastMessage = 'docs(security): update readme and API contracts for security';
+  final DateTime timestamp = DateTime.parse('2026-04-12T16:54:14+05:00');
 }